refactor: contact us page
