+++ components/contact-us/discord/online.vue
@@ -1,6 +1,6 @@
 <template>
   <iframe
-    :title="uniqueTitle"
+    title="COPS Insights"
     :src="`https://discord.com/widget?id=${serverid}&theme=dark`"
     allowtransparency="true"
     frameborder="0"
@@ -9,9 +9,16 @@
     height="500px"
   ></iframe>
 </template>
-<script lang="ts">
-export default {
+<script lang="ts">import { defineComponent } from '@nuxtjs/composition-api'
+
+export default defineComponent({
   // eslint-disable-next-line vue/require-prop-types
-  props: ['serverid']
-}
+  props: {
+    serverid: {
+      type: String,
+      default: '',
+      required: true
+    }
+  }
+})
 </script>
