+++ components/contact-us/map.vue
@@ -1,12 +1,11 @@
 <template>
   <iframe
-    :title="uniqueTitle"
+    title="IIT BHU Map"
     src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d3608.245385337551!2d82.9871848146768!3d25.26232953525508!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x398e33ef9b1f7cdd%3A0x7555db6d623dc140!2sIndian%20Institute%20of%20Technology%20(BHU)%20Varanasi!5e0!3m2!1sen!2sin!4v1634069543926!5m2!1sen!2sin"
     width="100%"
     height="400px"
-    allowfullscreen=""
+    allowfullscreen="false"
     loading="lazy"
     class="filter invert-90 hue-rotate-180 brightness-95 saturate-200"
   ></iframe>
 </template>
-<script lang="ts"></script>
